Adding http proxy to docker compose configuration
