+++ system-test/src/test/java/org/apache/mesos/elasticsearch/systemtest/TestBase.java
@@ -5,16 +5,18 @@ import com.github.dockerjava.api.command.InspectContainerResponse;
 import com.github.dockerjava.core.DockerClientBuilder;
 import com.github.dockerjava.core.DockerClientConfig;
 import com.mashape.unirest.http.Unirest;
-import org.apache.http.impl.client.HttpClients;
+import org.apache.http.HttpHost;
+import org.apache.log4j.Logger;
 import org.junit.Before;
 
+/**
+ * Base test class for everything shared between test cases
+ */
 public abstract class TestBase {
-    protected String getSlaveIp(String slaveName) {
-        DockerClientConfig config = DockerClientConfig.createDefaultConfigBuilder()
-                .withVersion("1.16")
-                .build();
+    private static final Logger LOGGER = Logger.getLogger(TestBase.class);
+    DockerClient docker;
 
-        DockerClient docker = DockerClientBuilder.getInstance(config).build();
+    protected String getSlaveIp(String slaveName) {
 
         InspectContainerResponse response = docker.inspectContainerCmd(slaveName).exec();
 
@@ -23,8 +25,15 @@ public abstract class TestBase {
 
     @Before
     public void setUp() throws Exception {
-        //set up http proxy if http.proxyHost and http.proxyPort system properties are defined
-        Unirest.setHttpClient(HttpClients.createSystem());
+        DockerClientConfig config = DockerClientConfig.createDefaultConfigBuilder()
+                .withVersion("1.16")
+                .build();
+
+        docker = DockerClientBuilder.getInstance(config).build();
+        HttpHost proxy = new HttpHost(config.getUri().getHost(), 8888);
+        Unirest.setProxy(proxy);
+        LOGGER.info("Using proxy " + proxy.toHostString());
+        System.out.println("proxy.toHostString() = " + proxy.toHostString());
         Unirest.setDefaultHeader("Content-Type", "application/json");
     }
 }
